Trying to fix bug with type int.

+++ cogs/destiny_api_helper_cogs.py
@@ -987,7 +987,7 @@ class destiny_api_helper_cogs(commands.Cog, name='Destiny Utilities'):
         costs = []
         for row in temp_item_df.itertuples(index=False):
             # calculate cost and append to list
-            cost = high_values[row.itemSubType.astype(int)] - row.desired_total.astype(int)
+            cost = high_values[row.itemSubType.astype(int)] - row.desired_total
             print(f'Reference points: {high_values[row.itemSubType.astype(int)]}, DF points: {row.desired_total.astype(int)}, Cost: {cost}')
             costs.append(cost)
